+++ dark-matter-analytics-mixpanel/src/main/java/me/melontini/dark_matter/impl/analytics/mixpanel/package-info.java
@@ -1,4 +1,0 @@
-@ApiStatus.Internal
-package me.melontini.dark_matter.impl.analytics.mixpanel;
-
-import org.jetbrains.annotations.ApiStatus;