+++ dark-matter-base/src/main/java/me/melontini/dark_matter/api/base/util/classes/Context.java
@@ -1,12 +1,15 @@
 package me.melontini.dark_matter.api.base.util.classes;
 
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
+import java.util.function.BiConsumer;
 
 public interface Context {
 
     <T> Optional<T> get(Class<T> cls, String key);
+    void forEach(BiConsumer<String, Object> consumer);
     default <T> T orThrow(Class<T> cls, String key) {
         return get(cls, key).orElseThrow(() -> new IllegalStateException("Missing required context '%s'!".formatted(key)));
     }
@@ -16,6 +19,14 @@ public interface Context {
             @Override
             public <T> Optional<T> get(Class<T> cls, String key) {
                 return Optional.empty();
+            }
+
+            @Override
+            public void forEach(BiConsumer<String, Object> consumer) {}
+
+            @Override
+            public String toString() {
+                return Collections.emptyMap().toString();
             }
         };
     }
@@ -27,6 +38,33 @@ public interface Context {
             public <T> Optional<T> get(Class<T> cls, String key) {
                 return Optional.ofNullable(cls.cast(ctx.get(key)));
             }
+
+            @Override
+            public void forEach(BiConsumer<String, Object> consumer) {
+                ctx.forEach(consumer);
+            }
+
+            @Override
+            public String toString() {
+                return ctx.toString();
+            }
         };
+    }
+
+    static Builder builder() {
+        return new Builder();
+    }
+
+    class Builder {
+        private final Map<String, Object> map = Collections.synchronizedMap(new HashMap<>());
+        private Builder() {}
+
+        public Builder put(String key, Object value) {
+            map.put(key, value);
+            return this;
+        }
+        public Context build() {
+            return Context.of(map);
+        }
     }
 }