+++ dark-matter-recipe-book/src/main/java/me/melontini/dark_matter/api/recipe_book/interfaces/PaginatedRecipeBookWidget.java
@@ -33,30 +33,4 @@ public interface PaginatedRecipeBookWidget {
     default int dm$getPageCount() {
         throw new IllegalStateException("Interface not implemented");
     }
-
-    @Deprecated(forRemoval = true)
-    default void updatePages() {
-        dm$updatePages();
-    }
-
-    @Deprecated(forRemoval = true)
-    default void updatePageSwitchButtons() {
-        dm$updatePageSwitchButtons();
-    }
-
-    @Deprecated(forRemoval = true)
-    default int getPage() {
-        return dm$getPage();
-    }
-
-    @Deprecated(forRemoval = true)
-    default void setPage(int page) {
-        dm$setPage(page);
-    }
-
-    @Deprecated(forRemoval = true)
-    default int getPageCount() {
-        return dm$getPageCount();
-    }
-
 }